+++ packages/ui/src/components/card.tsx
@@ -1,85 +1,84 @@
-import { ComponentProps, FC, ReactNode } from 'react'
-import { cn } from '@/lib/utils'
+import * as React from 'react'
+import { cn } from '@/lib/utils/index'
 
-type CardProps = ComponentProps<'div'> & {
-  aside?: ReactNode
+function Card({
+  className,
+  children,
+  aside,
+  ...props
+}: React.ComponentProps<'div'> & { aside?: React.ReactNode }) {
+  return (
+    <div
+      data-slot="card"
+      className={cn(
+        'bg-card text-card-foreground border-table-stroke flex flex-nowrap gap-2 rounded-xl border px-3 py-2',
+        className
+      )}
+      {...props}
+    >
+      {aside && <aside className="shrink-0">{aside}</aside>}
+      <div className="flex flex-1 flex-col gap-2">{children}</div>
+    </div>
+  )
 }
 
-export const Card: FC<CardProps> = ({ className, children, aside, ...props }) => (
-  <div
-    data-slot="card"
-    className={cn(
-      'flex flex-row gap-2 overflow-hidden rounded-sm border border-stone-100 bg-stone-100 px-2 py-3 text-sm tracking-tight text-black shadow-sm dark:border-black dark:bg-stone-900 dark:text-white dark:shadow-none',
-      className
-    )}
-    {...props}
-  >
-    <div className="order-2 flex flex-1 flex-col gap-2">{children}</div>
-    {aside && (
-      <aside data-slot="card-aside" className="order-1 max-w-8 shrink-0 overflow-hidden">
-        {aside}
-      </aside>
-    )}
-  </div>
-)
+function CardHeader({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      data-slot="card-header"
+      className={cn(
+        '@container/card-header grid auto-rows-min grid-rows-[auto_auto] items-start gap-0.5 has-data-[slot=card-action]:grid-cols-[1fr_auto] [.border-b]:pb-6',
+        className
+      )}
+      {...props}
+    />
+  )
+}
 
-type CardAsideCounterProps = ComponentProps<'div'>
-export const CardAsideCounter: FC<CardAsideCounterProps> = ({ className, ...props }) => (
-  <div
-    data-slot="card-aside-counter"
-    className={cn(
-      'min-h-5 min-w-5 overflow-hidden rounded-full border border-emerald-700 bg-emerald-700 px-0.5 pt-0.5 pb-1 text-center text-xs leading-none tracking-tight text-ellipsis whitespace-nowrap text-white dark:border-emerald-400 dark:bg-transparent dark:text-emerald-400',
-      className
-    )}
-    {...props}
-  />
-)
+function CardTitle({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      data-slot="card-title"
+      className={cn('text-foreground text-sm leading-none tracking-[-0.00875rem]', className)}
+      {...props}
+    />
+  )
+}
 
-type CardHeaderProps = ComponentProps<'div'>
-export const CardHeader: FC<CardHeaderProps> = ({ className, ...props }) => (
-  <div data-slot="card-header" className={cn('flex flex-col gap-0.5', className)} {...props} />
-)
+function CardDescription({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      data-slot="card-description"
+      className={cn('text-muted text-xs leading-none tracking-[-0.00875rem]', className)}
+      {...props}
+    />
+  )
+}
 
-type CardTitleProps = ComponentProps<'div'>
-export const CardTitle: FC<CardTitleProps> = ({ className, ...props }) => (
-  <div
-    data-slot="card-title"
-    className={cn('text-sm leading-none tracking-tight text-black dark:text-white', className)}
-    {...props}
-  />
-)
+function CardAction({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      data-slot="card-action"
+      className={cn('col-start-2 row-span-2 row-start-1 self-start justify-self-end', className)}
+      {...props}
+    />
+  )
+}
 
-type CardDescriptionProps = ComponentProps<'div'>
-export const CardDescription: FC<CardDescriptionProps> = ({ className, ...props }) => (
-  <div
-    data-slot="card-description"
-    className={cn(
-      'text-xs leading-none tracking-tight text-black/50 dark:text-white/50',
-      className
-    )}
-    {...props}
-  />
-)
+function CardContent({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div data-slot="card-content" className={cn('text-foreground text-sm', className)} {...props} />
+  )
+}
 
-type CardContentProps = ComponentProps<'div'>
-export const CardContent: FC<CardContentProps> = ({ className, ...props }) => (
-  <div
-    data-slot="card-content"
-    className={cn('text-base leading-none tracking-tight text-black dark:text-white', className)}
-    {...props}
-  />
-)
+function CardFooter({ className, ...props }: React.ComponentProps<'div'>) {
+  return (
+    <div
+      data-slot="card-footer"
+      className={cn('text-foreground flex items-center text-sm [.border-t]:pt-6', className)}
+      {...props}
+    />
+  )
+}
 
-type CardActionProps = ComponentProps<'div'>
-export const CardAction: FC<CardActionProps> = ({ className, ...props }) => (
-  <div
-    data-slot="card-action"
-    className={cn('flex flex-row-reverse gap-4', className)}
-    {...props}
-  />
-)
-
-type CardFooterProps = ComponentProps<'div'>
-export const CardFooter: FC<CardFooterProps> = ({ className, ...props }) => (
-  <div data-slot="card-footer" className={cn('pt-4', className)} {...props} />
-)
+export { Card, CardHeader, CardFooter, CardTitle, CardAction, CardDescription, CardContent }